+++ src/components/CalendarPanel/CalendarPanel.test.tsx
@@ -1,7 +1,18 @@
-import { shallow } from 'enzyme';
 import React from 'react';
 import { toDataFrame } from '@grafana/data';
+import { getBackendSrv } from '@grafana/runtime';
+import { act, render, screen, waitFor } from '@testing-library/react';
+import { TestIds } from '../../constants';
 import { CalendarPanel } from './CalendarPanel';
+
+/**
+ * Mock @grafana/runtime
+ */
+jest.mock('@grafana/runtime', () => ({
+  getBackendSrv: jest.fn(() => ({
+    get: jest.fn(() => Promise.resolve()),
+  })),
+}));
 
 /**
  * Panel
@@ -22,8 +33,24 @@ describe('Panel', () => {
       );
     };
 
-    const wrapper = shallow(getComponent({}));
-    const div = wrapper.find('div');
-    expect(div.exists()).toBeTruthy();
+    const getMock = jest.fn(() => Promise.resolve());
+    jest.mocked(getBackendSrv).mockImplementationOnce(
+      () =>
+        ({
+          get: getMock,
+        } as any)
+    );
+
+    await act(async () => {
+      await render(getComponent({}));
+
+      /**
+       * Remove act warnings
+       * Wait timeout until promise resolves. Because there is not loading element for checking promise resolves
+       */
+      await new Promise((resolve) => setTimeout(resolve, 1));
+    });
+
+    await waitFor(() => expect(screen.getByTestId(TestIds.panel.root)).toBeInTheDocument());
   });
 });