+++ src/components/Day/Day.test.tsx
@@ -1,6 +1,7 @@
 import dayjs from 'dayjs';
-import { shallow } from 'enzyme';
 import React from 'react';
+import { render, screen } from '@testing-library/react';
+import { TestIds } from '../../constants';
 import { Day } from './Day';
 
 /**
@@ -12,10 +13,8 @@ describe('Day', () => {
       return <Day {...restProps} />;
     };
 
-    const wrapper = shallow(
-      getComponent({ day: dayjs(Date.now()), from: dayjs(Date.now()), to: dayjs(Date.now()), events: [] })
-    );
-    const div = wrapper.find('div');
-    expect(div.exists()).toBeTruthy();
+    render(getComponent({ day: dayjs(Date.now()), from: dayjs(Date.now()), to: dayjs(Date.now()), events: [] }));
+
+    expect(screen.getByTestId(TestIds.day.root)).toBeInTheDocument();
   });
 });