+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Rebuild using Grafana 9.4.7 (#96)
 - Add Text Formatting, Mappings and Overrides (#97)
 - Remove unused Standard Options (#97)
+- Increase Test Coverage and update testing library (#100)
 
 ### Bug fixes